feat(ux): extract BackToTopButton and add to about and lens detail pages
Extract the scroll-to-top button from LensListClient into a reusable
BackToTopButton component, then add it to the About page and the lens
detail page where long content makes it useful.
