+++ src/app/[locale]/about/page.tsx
@@ -1,6 +1,7 @@
 import type { Metadata } from "next";
 import { getTranslations } from "next-intl/server";
 import AboutContent from "@/components/AboutContent";
+import BackToTopButton from "@/components/BackToTopButton";
 
 export async function generateMetadata(): Promise<Metadata> {
   const t = await getTranslations("About");
@@ -9,6 +10,9 @@ export async function generateMetadata(): Promise<Metadata> {
 
 export default function AboutPage() {
   return (
-    <AboutContent />
+    <>
+      <AboutContent />
+      <BackToTopButton />
+    </>
   );
 }
